support send of True and False OSC 1.1 types
the paramaters list still requires an element to be present for the True or False type, though the actual value is unimportant, it could be null or undefined since the type is the value of sorts - this was done for symmetry to the reader which expands the T and F types to true and false paramaters.

alternatively, the params could exclude that item and makeMessageOctets() could params.shift() as the type and paramater lists would not be of equal length.