synthesize Impulse param as null instead of Infinity
until a more appropraite value can be determined use null as Infinity was a OSC 1.0 type reference. the only problem with this is the receiver now needs to parse the type tags now on null to determine if a Null or Impulse is received -- perhaps type tags should be an array?